refactor: move config commands to per-command file in config dir
